Added a "mkfs <label>" command.
Note:

No checking is performed on the label, so this command may
be used to effect a low-level format on a IEC drive, but be
aware that this takes long time. The CREATED/ERROR events
are pre-allocated, so they will get through, but clock
events will fill the queue before the format finishes, so
other events will be lost.